clang でコンパイルできるように明示的にキャスト。
clang 4.0.0 で > 演算子のオーバーロードが解決できなかった。

+++ core/common/jrpc.cpp
@@ -120,7 +120,7 @@ json JrpcApi::getLog(json::array_t args)
 
     if (maxLines == nullptr)
         maxLines = lines.size();
-    while (lines.size() > maxLines)
+    while (lines.size() > (size_t)maxLines)
         lines.pop_back();
 
     auto log = str::join("\n", lines);
